pyproject.toml
[tool.poetry]
name = "course-aplication"
version = "0.1.0"
description = ""
authors = ["diana"]
readme = "README.md"
packages = [{include = "course_aplication"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic-cli = "^4.3.0"
uvicorn = "^0.20.0"
fastapi = "^0.92.0"
enum = "^0.4.7"
DateTime = "^5.0"
typing = "^3.10.0.0"
databases = "^0.7.0"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
